[#222] fix: 포스트 갤러리 - 포스트 카드 연동
